update component to standalone
