+++ libs/design/select/examples/src/disabled-select/disabled-select.component.html
@@ -1,4 +1,4 @@
-<daff-select [options]="options" [formControl]="disabled" [disabled]="true">
+<daff-select [options]="options" [formControl]="disabled">
   <label daffFormLabel>Select an address</label>
   <ng-template daffSelectOption let-option="option"></ng-template>
 </daff-select>
