+++ libs/design/select/examples/src/disabled-select/disabled-select.component.ts
@@ -2,7 +2,12 @@ import {
   ChangeDetectionStrategy,
   Component,
 } from '@angular/core';
-import { UntypedFormControl } from '@angular/forms';
+import {
+  ReactiveFormsModule,
+  UntypedFormControl,
+} from '@angular/forms';
+
+import { DAFF_SELECT_COMPONENTS } from '@daffodil/design/select';
 
 import { SELECT_EXAMPLE_ADDRESSES } from '../models/addresses';
 
@@ -11,6 +16,11 @@ import { SELECT_EXAMPLE_ADDRESSES } from '../models/addresses';
   selector: 'disabled-select',
   templateUrl: './disabled-select.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [
+    ReactiveFormsModule,
+    DAFF_SELECT_COMPONENTS,
+  ],
 })
 export class DisabledSelectComponent {
   disabled = new UntypedFormControl({ value : '' , disabled: true });
