--- libs/design/select/examples/src/examples.ts
@@ -1,11 +1,0 @@
-import { DefaultSelectComponent } from './default-select/default-select.component';
-import { DisabledSelectComponent } from './disabled-select/disabled-select.component';
-import { SelectWithErrorComponent } from './select-with-error/select-with-error.component';
-import { SkeletonSelectComponent } from './skeleton-select/skeleton-select.component';
-
-export const SELECT_EXAMPLES = [
-  DefaultSelectComponent,
-  DisabledSelectComponent,
-  SkeletonSelectComponent,
-  SelectWithErrorComponent,
-];
