+++ libs/design/select/examples/src/public_api.ts
@@ -1,9 +1,11 @@
-export { DefaultSelectComponent } from './default-select/default-select.component';
-export { DefaultSelectModule } from './default-select/default-select.module';
-export { DisabledSelectComponent } from './disabled-select/disabled-select.component';
-export { DisabledSelectModule } from './disabled-select/disabled-select.module';
-export { SkeletonSelectComponent } from './skeleton-select/skeleton-select.component';
-export { SkeletonSelectModule } from './skeleton-select/skeleton-select.module';
-export { SelectWithErrorModule } from './select-with-error/select-with-error.module';
-export { SelectWithErrorComponent } from './select-with-error/select-with-error.component';
-export { SELECT_EXAMPLES } from './examples';
+import { DefaultSelectComponent } from './default-select/default-select.component';
+import { DisabledSelectComponent } from './disabled-select/disabled-select.component';
+import { SelectWithErrorComponent } from './select-with-error/select-with-error.component';
+import { SkeletonSelectComponent } from './skeleton-select/skeleton-select.component';
+
+export const SELECT_EXAMPLES = [
+  DefaultSelectComponent,
+  DisabledSelectComponent,
+  SkeletonSelectComponent,
+  SelectWithErrorComponent,
+];
