--- libs/design/select/examples/src/skeleton-select/skeleton-select.module.ts
@@ -1,22 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { ReactiveFormsModule } from '@angular/forms';
-
-import { DaffSelectModule } from '@daffodil/design/select';
-
-import { SkeletonSelectComponent } from './skeleton-select.component';
-
-@NgModule({
-  declarations: [
-    SkeletonSelectComponent,
-  ],
-  exports: [
-    SkeletonSelectComponent,
-  ],
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    DaffSelectModule,
-  ],
-})
-export class SkeletonSelectModule { }
