+++ libs/design/select/ng-package.json
@@ -1,9 +1,7 @@
 {
-  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/design/select",
-  "deleteDestPath": false,
+  "$schema": "../../../node_modules/ng-packagr/ng-entrypoint.schema.json",
   "lib": {
     "entryFile": "src/index.ts",
     "styleIncludePaths": ["../src/scss"]
   }
-} 
+}
