+++ libs/design/select/src/option/option.directive.ts
@@ -9,6 +9,7 @@ import {
  */
 @Directive({
   selector: '[daffSelectOption]',
+  standalone: true,
 })
 export class DaffSelectOptionDirective {
   constructor(@Optional() public templateRef?: TemplateRef<unknown>) {}
