+++ libs/design/select/src/public_api.ts
@@ -2,3 +2,4 @@ export { DaffSelectModule } from './select.module';
 export * from './select/select.component';
 export * from './option/option.directive';
 export * from './option/context.type';
+export { DAFF_SELECT_COMPONENTS } from './select';
