+++ libs/design/select/src/select/select.component.html
@@ -18,7 +18,6 @@
       [ngTemplateOutletContext]="{option: value}"><div></div>
     </ng-container>
   </div>
-  <fa-icon daffSuffix [icon]="faChevronDown"></fa-icon>
 </button>
 <ng-template cdkPortal #optionsTemplate="cdkPortal">
   <div class="daff-select__options" #optionsEl
