+++ libs/design/select/src/select/select.component.scss
@@ -1,4 +1,3 @@
-@use '@angular/cdk/overlay-prebuilt';
 @use '../../../scss/interactions';
 @use '../../../scss/state';
 
@@ -11,17 +10,18 @@
 		transition: transform 150ms;
 	}
 
-	&#{$root}--open {
+	&.open {
 		#{$root}__field {
 			border-radius: 4px 4px 0 0;
-		}
 
-		.daff-suffix {
-			transform: rotate(180deg);
+			&:after {
+				top: 56%;
+				transform: translateY(-50%) rotate(225deg);
+			}
 		}
 	}
 
-	&#{$root}--disabled {
+	&.disabled {
 		#{$root}__field {
 			cursor: not-allowed;
 		}
@@ -31,6 +31,10 @@
 		#{$root}__field {
 			@include state.skeleton-screen(100%, 54px);
 			border: none;
+
+			&::after {
+				visibility: hidden;
+			}
 
 			> * {
 				visibility: hidden;
@@ -51,6 +55,21 @@
 		padding: 16px;
 		text-align: left;
 		width: 100%;
+		position: relative;
+
+		&::after {
+      content: '';
+      position: absolute;
+      top: 48%;
+      right: 1rem;
+      display: inline-block;
+      border-right: 2px solid currentColor;
+      border-bottom: 2px solid currentColor;
+      width: 0.5rem;
+      height: 0.5rem;
+      transform: translateY(-50%) rotate(45deg);
+      transition: transform 150ms;
+    }
 	}
 
 	.daff-form-label {
